Change access modifiers accordingly
