--- sdks/java/io/cdc/src/main/java/org/apache/beam/io/cdc/DebeziumBeamSourceTaskContext.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.beam.io.cdc;
-
-import org.apache.kafka.connect.source.SourceTaskContext;
-import org.apache.kafka.connect.storage.OffsetStorageReader;
-import org.checkerframework.checker.nullness.qual.Nullable;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Map;
-
-@SuppressWarnings({"nullness"})
-public class DebeziumBeamSourceTaskContext implements SourceTaskContext {
-    private static final Logger LOG = LoggerFactory.getLogger(DebeziumBeamSourceTaskContext.class);
-
-    private final @Nullable Map<String, ?> initialOffset;
-
-    DebeziumBeamSourceTaskContext(@Nullable Map<String, ?> initialOffset) {
-        this.initialOffset = initialOffset;
-    }
-
-    @Override
-    public Map<String, String> configs() {
-        // TODO(pabloem): Do we need to implement this?
-        throw new UnsupportedOperationException("unimplemented");
-    }
-
-    @Override
-    public OffsetStorageReader offsetStorageReader() {
-        LOG.debug("------------- Creating an offset storage reader");
-        return new DebeziumSourceOffsetStorageReader(initialOffset);
-    }
-}
