--- sdks/java/io/cdc/src/main/java/org/apache/beam/io/cdc/DebeziumOffsetHolder.java
@@ -1,38 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.beam.io.cdc;
-
-import org.checkerframework.checker.nullness.qual.Nullable;
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
-public class DebeziumOffsetHolder implements Serializable {
-    public final @Nullable Map<String, ?> offset;
-    public final @Nullable List<?> history;
-    public final @Nullable Integer fetchedRecords;
-
-    DebeziumOffsetHolder(@Nullable Map<String, ?> offset,
-                         @Nullable List<?> history,
-                         @Nullable Integer fetchedRecords) {
-        this.offset = offset;
-        this.history = history == null ? new ArrayList<>() : history;
-        this.fetchedRecords = fetchedRecords;
-    }
-}
