+++ sdks/java/io/cdc/src/main/java/org/apache/beam/io/cdc/KafkaSourceConsumerFn.java
@@ -17,29 +17,40 @@
  */
 package org.apache.beam.io.cdc;
 
+import io.debezium.document.Document;
+import io.debezium.document.DocumentReader;
+import io.debezium.document.DocumentWriter;
+import io.debezium.relational.history.AbstractDatabaseHistory;
+import io.debezium.relational.history.DatabaseHistoryException;
+import io.debezium.relational.history.HistoryRecord;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.SerializableCoder;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.splittabledofn.RestrictionTracker;
+import org.apache.beam.sdk.transforms.splittabledofn.SplitResult;
+import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableMap;
 import org.apache.kafka.connect.source.SourceConnector;
 import org.apache.kafka.connect.source.SourceRecord;
 import org.apache.kafka.connect.source.SourceTask;
+import org.apache.kafka.connect.source.SourceTaskContext;
+import org.apache.kafka.connect.storage.OffsetStorageReader;
+import org.checkerframework.checker.nullness.qual.Nullable;
 import org.joda.time.DateTime;
 import org.joda.time.Duration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.io.Serializable;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Consumer;
 
 /**
  * <h3>Quick Overview</h3>
  * SDF used to process records fetched from supported Debezium Connectors.
  *
- * Currently it has a time limiter (see {@link DebeziumOffsetTracker}) which, if set,
+ * Currently it has a time limiter (see {@link OffsetTracker}) which, if set,
  * it will stop automatically after the specified elapsed minutes. Otherwise, it will keep
  * running until the user explicitly interrupts it.
  *
@@ -59,7 +70,7 @@ public class KafkaSourceConsumerFn<T> extends DoFn<Map<String, String>, T> {
 
     private final Class<? extends SourceConnector> connectorClass;
     private final SourceRecordMapper<T> fn;
-    protected static final Map<String, RestrictionTracker<DebeziumOffsetHolder,  Map<String, Object>>>
+    protected static final Map<String, RestrictionTracker<OffsetHolder,  Map<String, Object>>>
     restrictionTrackers = new ConcurrentHashMap<>();
 
     /**
@@ -86,19 +97,19 @@ public class KafkaSourceConsumerFn<T> extends DoFn<Map<String, String>, T> {
     }
 
     @GetInitialRestriction
-    public DebeziumOffsetHolder getInitialRestriction(@Element Map<String, String> unused) throws IOException {
+    public OffsetHolder getInitialRestriction(@Element Map<String, String> unused) throws IOException {
         KafkaSourceConsumerFn.startTime = new DateTime();
-        return new DebeziumOffsetHolder(null, null, null);
+        return new OffsetHolder(null, null, null);
     }
 
     @NewTracker
-    public RestrictionTracker<DebeziumOffsetHolder, Map<String, Object>> newTracker(@Restriction DebeziumOffsetHolder restriction) {
-        return new DebeziumOffsetTracker(restriction);
+    public RestrictionTracker<OffsetHolder, Map<String, Object>> newTracker(@Restriction OffsetHolder restriction) {
+        return new OffsetTracker(restriction);
     }
 
     @GetRestrictionCoder
-    public Coder<DebeziumOffsetHolder> getRestrictionCoder() {
-        return SerializableCoder.of(DebeziumOffsetHolder.class);
+    public Coder<OffsetHolder> getRestrictionCoder() {
+        return SerializableCoder.of(OffsetHolder.class);
     }
 
     /**
@@ -111,7 +122,7 @@ public class KafkaSourceConsumerFn<T> extends DoFn<Map<String, String>, T> {
      */
     @DoFn.ProcessElement
     public ProcessContinuation process(@Element Map<String, String> element,
-                                       RestrictionTracker<DebeziumOffsetHolder,
+                                       RestrictionTracker<OffsetHolder,
                                                Map<String, Object>> tracker,
                                        OutputReceiver<T> receiver) throws Exception {
         Map<String, String> configuration = new HashMap<>(element);
@@ -125,7 +136,7 @@ public class KafkaSourceConsumerFn<T> extends DoFn<Map<String, String>, T> {
 
         SourceTask task = (SourceTask) connector.taskClass().getDeclaredConstructor().newInstance();
 
-        task.initialize(new DebeziumBeamSourceTaskContext(tracker.currentRestriction().offset));
+        task.initialize(new BeamSourceTaskContext(tracker.currentRestriction().offset));
         task.start(connector.taskConfigs(1).get(0));
 
         List<SourceRecord> records = task.poll();
@@ -165,5 +176,188 @@ public class KafkaSourceConsumerFn<T> extends DoFn<Map<String, String>, T> {
 
     public String getHashCode() {
         return Integer.toString(System.identityHashCode(this));
+    }
+
+    private static class BeamSourceTaskContext implements SourceTaskContext {
+        private final @Nullable Map<String, ?> initialOffset;
+
+        BeamSourceTaskContext(@Nullable Map<String, ?> initialOffset) {
+            this.initialOffset = initialOffset;
+        }
+
+        @Override
+        public Map<String, String> configs() {
+            // TODO(pabloem): Do we need to implement this?
+            throw new UnsupportedOperationException("unimplemented");
+        }
+
+        @Override
+        public OffsetStorageReader offsetStorageReader() {
+            LOG.debug("------------- Creating an offset storage reader");
+            return new DebeziumSourceOffsetStorageReader(initialOffset);
+        }
+    }
+
+    private static class DebeziumSourceOffsetStorageReader implements OffsetStorageReader {
+        private final Map<String, ?> offset;
+
+        DebeziumSourceOffsetStorageReader(Map<String, ?> initialOffset) {
+            this.offset = initialOffset;
+        }
+
+        @Override
+        public <V> Map<String, Object> offset(Map<String, V> partition) {
+            return offsets(Collections.singletonList(partition)).getOrDefault(partition, ImmutableMap.of());
+        }
+
+        @Override
+        public <T> Map<Map<String, T>, Map<String, Object>> offsets(Collection<Map<String, T>> partitions) {
+            LOG.debug("-------------- GETTING OFFSETS!");
+
+            Map<Map<String, T>, Map<String, Object>> map = new HashMap<>();
+            for (Map<String, T> partition : partitions) {
+                map.put(partition, (Map<String, Object>) offset);
+            }
+
+            LOG.debug("-------------- OFFSETS: {}", map);
+            return map;
+        }
+    }
+
+    static class OffsetHolder implements Serializable {
+        public final @Nullable Map<String, ?> offset;
+        public final @Nullable List<?> history;
+        public final @Nullable Integer fetchedRecords;
+
+        OffsetHolder(@Nullable Map<String, ?> offset,
+                     @Nullable List<?> history,
+                     @Nullable Integer fetchedRecords) {
+            this.offset = offset;
+            this.history = history == null ? new ArrayList<>() : history;
+            this.fetchedRecords = fetchedRecords;
+        }
+    }
+
+    /**
+     * {@link RestrictionTracker} for Debezium connectors
+     */
+    static class OffsetTracker extends RestrictionTracker<OffsetHolder, Map<String, Object>> {
+        private OffsetHolder restriction;
+        private static final long MILLIS = 60 * 1000;
+
+        OffsetTracker(OffsetHolder holder) {
+            this.restriction = holder;
+        }
+
+        /**
+         * Overriding {@link #tryClaim} in order to stop fetching records from the database.
+         *
+         * <p>This works on two different ways:</p>
+         * <h3>Number of records</h3>
+         * <p>
+         *     This is the default behavior.
+         *     Once the specified number of records has been reached, it will stop fetching them.
+         * </p>
+         * <h3>Time based</h3>
+         * </p>
+         *     User may specify the amount of time the connector to be kept alive.
+         *     Please see {@link KafkaSourceConsumerFn} for more details on this.
+         * </p>
+         *
+         *
+         * @param position Currently not used
+         * @return boolean
+         */
+        @Override
+        public boolean tryClaim(Map<String, Object> position) {
+            LOG.debug("-------------- Claiming {} used to have: {}", position, restriction.offset);
+            long elapsedTime = System.currentTimeMillis() - startTime.getMillis();
+            int fetchedRecords = this.restriction.fetchedRecords == null ? 0 : this.restriction.fetchedRecords + 1;
+            LOG.debug("-------------- Time running: {} / {}", elapsedTime, (minutesToRun * MILLIS));
+            this.restriction = new OffsetHolder(position, this.restriction.history, fetchedRecords);
+            LOG.debug("-------------- History: {}", this.restriction.history);
+
+            if (maxRecords == null && minutesToRun == -1) {
+                return true;
+            }
+
+            if (maxRecords != null && minutesToRun < 0) {
+                return fetchedRecords < maxRecords;
+            }
+
+            return elapsedTime < minutesToRun * MILLIS;
+        }
+
+        @Override
+        public OffsetHolder currentRestriction() {
+            return restriction;
+        }
+
+        @Override
+        public SplitResult<OffsetHolder> trySplit(double fractionOfRemainder) {
+            LOG.debug("-------------- Trying to split: fractionOfRemainder={}", fractionOfRemainder);
+            return SplitResult.of(new OffsetHolder(null, null, null), restriction);
+        }
+
+        @Override
+        public void checkDone() throws IllegalStateException {
+        }
+
+        @Override
+        public IsBounded isBounded() {
+            return IsBounded.BOUNDED;
+        }
+    }
+
+    public static class DebeziumSDFDatabaseHistory extends AbstractDatabaseHistory {
+        private List<byte[]> history;
+
+        public DebeziumSDFDatabaseHistory() {
+            this.history = new ArrayList<>();
+        }
+
+        @Override
+        public void start() {
+            super.start();
+            LOG.debug("------------ STARTING THE DATABASE HISTORY! - trackers: {} - config: {}",
+                    restrictionTrackers, config.asMap());
+
+            // TODO(pabloem): Figure out how to link these. For now, we'll just link directly.
+            RestrictionTracker<OffsetHolder, ?> tracker = restrictionTrackers.get(
+                    // JUST GETTING THE FIRST KEY. This will not work in the future.
+                    restrictionTrackers.keySet().iterator().next());
+            this.history = (List<byte[]>) tracker.currentRestriction().history;
+        }
+
+        @Override
+        protected void storeRecord(HistoryRecord record) throws DatabaseHistoryException {
+            LOG.debug("------------- Adding history! {}", record);
+
+            history.add(DocumentWriter.defaultWriter().writeAsBytes(record.document()));
+        }
+
+        @Override
+        protected void recoverRecords(Consumer<HistoryRecord> consumer) {
+            LOG.debug("------------- Trying to recover!");
+
+            try {
+                for (byte[] record : history) {
+                    Document doc = DocumentReader.defaultReader().read(record);
+                    consumer.accept(new HistoryRecord(doc));
+                }
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
+
+        @Override
+        public boolean exists() {
+            return history != null && !history.isEmpty();
+        }
+
+        @Override
+        public boolean storageExists() {
+            return history != null && !history.isEmpty();
+        }
     }
 }
